pyproject.toml
[project]
name = "decelerates"
version = "0.1.0"
description = "Norwegian company due-diligence tool"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.129.0",
    "uvicorn[standard]==0.41.0",
    "sqlalchemy==2.0.46",
    "psycopg[binary]==3.3.3",
    "requests==2.33.1",
    "pydantic==2.12.5",
    "streamlit==1.54.0",
    "pandas==2.3.3",
    "anthropic==0.84.0",
    "pgvector==0.4.2",
    "voyageai==0.2.4",
    "google-genai==1.66.0",
    "pdfplumber==0.11.9",
    "fpdf2==2.8.7",
    "python-multipart==0.0.22",
    "playwright==1.58.0",
    "langchain-core==1.2.24",
    "langchain-text-splitters==1.1.1",
    "openai>=1.30.0",
    "azure-ai-documentintelligence>=1.0.0",
    "azure-storage-blob>=12.19.0",
    "azure-identity>=1.15.0",
    "azure-search-documents>=11.4.0",
    "azure-communication-email>=1.0.0",
    "pyasn1>=0.6.3",
    "aiohttp>=3.13.5",
    "pygments>=2.20.0",
    "cryptography>=46.0.6",
    "slowapi>=0.1.9",
    "alembic>=1.13.0",
    "punq>=0.7.0",
    "pyjwt[crypto]>=2.8.0",
    "azure-monitor-opentelemetry>=1.6.0",
    "openpyxl>=3.1.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pylint>=3.0.0",
    "radon>=6.0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["api", "ui"]

[tool.pytest.ini_options]
markers = [
    "system: end-to-end tests requiring a live API instance (set SYSTEM_TEST_URL)",
]


# ── Ruff configuration ────────────────────────────────────────────────────────
# Ignores documented pre-existing stylistic violations so ruff can actually
# run in CI. The external `reusable-ruff-formating.yml@v0.18.0` action reads
# this config. All ignored rules are intentional patterns, not latent bugs:
#
#   E402  — module-level import not at top: widely used for sys.modules stub
#           setup before imports (see tests/conftest.py) and conditional
#           Azure Monitor configuration in api/main.py.
#   E701  — multiple statements on colon: used in short Streamlit widget
#           blocks where inlining is more readable than 3-line spreads.
#   E702  — multiple statements on semicolon: same rationale.
#   E741  — ambiguous variable name (l, I, O): used in short loop variables
#           where the context makes the meaning unambiguous.
#
# F401 (unused imports) and F821 (undefined names) are NOT ignored — F821
# caught a real bug in ui/views/admin.py:_render_data_controls missing a
# `headers = get_auth_headers()` call.
[tool.ruff.lint]
ignore = ["E402", "E701", "E702", "E741"]
